backend/config/settings.py
import os
from dotenv import load_dotenv
from pydantic_settings import BaseSettings

# Load environment variables from .env file
load_dotenv()

class Settings(BaseSettings):
    gemini_api_key: str = os.getenv("GEMINI_API_KEY", "")
    environment: str = os.getenv("ENVIRONMENT", "development")
    log_level: str = os.getenv("LOG_LEVEL", "INFO")
    
    # Agent configuration
    max_response_tokens: int = 1000
    temperature: float = 0.7
    
    class Config:
        env_file = ".env"

# Global settings instance
settings = Settings() 